Drop-down menu for algorithms. Draw style changes.
Version 1 as stated in README.md

Implemented a new shuffle algorithm that looks prettier because it looks more "random". Implemented a drop-down menu for choosing algorithms.

Found a bug when using on the live version, the array won't show on mobile. Stackoverflow said it might be a "let" vs "var" problem. Will experiment with that option next version.